Fixing the widget duplication, ignore the three new themes i accidentally added them and couldn't get them removed. OH and we yeeted the Tkinter plans!
