+++ dataset_tools/themes/KTISEOS_NYX_THEMES/Watermelon_Unreadable.qss
@@ -1,0 +1,235 @@
+/* General Styles */
+QWidget {
+    background-color: #564946;
+    color: #69ebd0;
+    border: 1px solid #512888;
+}
+
+/* Buttons */
+QPushButton {
+    background-color: #ca4663;
+    color: #69ebd0;
+    border: 1px solid #512888;
+    padding: 5px;
+    border-radius: 3px;
+}
+
+QPushButton:hover {
+    background-color: #59e0b7;
+}
+
+QPushButton:pressed {
+    background-color: #512888;
+}
+
+/* LineEdit and TextEdit */
+QLineEdit, QTextEdit {
+    background-color: #59e0b7;
+    color: #564946;
+    border: 1px solid #512888;
+    padding: 3px;
+    border-radius: 3px;
+}
+
+/* Labels */
+QLabel {
+    color: #69ebd0;
+}
+
+/* GroupBox */
+QGroupBox {
+    background-color: #564946;
+    border: 1px solid #512888;
+    color: #69ebd0;
+    margin-top: 1ex; /* leave space at the top for the title */
+}
+
+QGroupBox::title {
+    subcontrol-origin: margin;
+    subcontrol-position: top left; /* position at the top left */
+    padding: 0 3px;
+    background-color: #ca4663;
+    color: #69ebd0;
+}
+
+/* ComboBox */
+QComboBox {
+    background-color: #59e0b7;
+    color: #564946;
+    border: 1px solid #512888;
+    padding: 1px 0px 1px 3px; /* top, right, bottom, left */
+    border-radius: 3px;
+}
+
+QComboBox::drop-down {
+    subcontrol-origin: padding;
+    subcontrol-position: top right;
+    width: 15px;
+    border-left-width: 1px;
+    border-left-color: #512888;
+    border-left-style: solid; /* just a single line */
+    border-top-right-radius: 3px; /* same radius as the QComboBox */
+    border-bottom-right-radius: 3px;
+}
+
+QComboBox::down-arrow {
+    image: url(down_arrow.png); /* Placeholder for an arrow icon */
+}
+
+/* ScrollBar */
+QScrollBar:vertical {
+    border: 1px solid #512888;
+    background: #564946;
+    width: 10px;
+    margin: 15px 0 15px 0;
+}
+
+QScrollBar::handle:vertical {
+    background: #ca4663;
+    min-height: 20px;
+}
+
+QScrollBar::add-line:vertical {
+    border: 1px solid #512888;
+    background: #59e0b7;
+    height: 15px;
+    subcontrol-origin: bottom;
+    subcontrol-position: bottom;
+}
+
+QScrollBar::sub-line:vertical {
+    border: 1px solid #512888;
+    background: #59e0b7;
+    height: 15px;
+    subcontrol-origin: top;
+    subcontrol-position: top;
+}
+
+QScrollBar::up-arrow:vertical, QScrollBar::down-arrow:vertical {
+    background: #69ebd0;
+}
+
+QScrollBar::add-page:vertical, QScrollBar::sub-page:vertical {
+    background: none;
+}
+
+/* CheckBox */
+QCheckBox {
+    color: #69ebd0;
+}
+
+QCheckBox::indicator {
+    width: 13px;
+    height: 13px;
+    border: 1px solid #512888;
+    background-color: #59e0b7;
+}
+
+QCheckBox::indicator:checked {
+    background-color: #ca4663;
+}
+
+/* RadioButton */
+QRadioButton {
+    color: #69ebd0;
+}
+
+QRadioButton::indicator {
+    width: 13px;
+    height: 13px;
+    border: 1px solid #512888;
+    border-radius: 6px; /* Makes it circular */
+    background-color: #59e0b7;
+}
+
+QRadioButton::indicator:checked {
+    background-color: #ca4663;
+}
+
+/* TabWidget */
+QTabWidget::pane {
+    border: 1px solid #512888;
+    background-color: #564946;
+}
+
+QTabBar::tab {
+    background: #59e0b7;
+    border: 1px solid #512888;
+    border-bottom-color: #512888; /* same as pane color */
+    border-top-left-radius: 4px;
+    border-top-right-radius: 4px;
+    min-width: 8ex;
+    padding: 2px;
+    color: #564946;
+}
+
+QTabBar::tab:selected, QTabBar::tab:hover {
+    background: #ca4663;
+}
+
+QTabBar::tab:selected {
+    border-color: #512888;
+    border-bottom-color: #ca4663; /* make the selected tab appear above the pane */
+}
+
+/* MenuBar */
+QMenuBar {
+    background-color: #59e0b7;
+    color: #564946;
+}
+
+QMenuBar::item {
+    spacing: 3px;
+    padding: 1px 4px;
+    background: transparent;
+    border-radius: 4px;
+}
+
+QMenuBar::item:selected {
+    background: #ca4663;
+}
+
+QMenuBar::item:pressed {
+    background: #512888;
+}
+
+/* Menu */
+QMenu {
+    background-color: #59e0b7;
+    border: 1px solid #512888;
+    border-radius: 3px;
+}
+
+QMenu::item {
+    padding: 2px 25px 2px 20px;
+    border: 1px solid transparent; /* reserve space for selection border */
+    color: #564946;
+}
+
+QMenu::item:selected {
+    background-color: #ca4663;
+    color: #69ebd0;
+}
+
+QMenu::separator {
+    height: 1px;
+    background: #512888;
+    margin-left: 10px;
+    margin-right: 10px;
+}
+
+QListWidget::item {
+    background-color: transparent;
+    color: #F2C7CA;  /* Light text - matches your main theme */
+}
+
+QListWidget::item:selected {
+    background-color: #A87A81;
+    color: #F2C7CA;
+}
+
+QListWidget {
+    background-color: #564946;
+    border: 1px solid #512888;
+    color: #69ebd0;
+}
